backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
google-auth-oauthlib==1.1.0
google-api-python-client==2.108.0
beautifulsoup4==4.12.2
lxml==4.9.3
cryptography==41.0.7
python-dotenv==1.0.0
python-multipart==0.0.6
email-validator==2.1.0
celery==5.3.4
redis==5.0.1
